backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
modal>=0.62.0
aiofiles==23.2.1
numpy>=1.24.0
